organize CRA src to neatly separate concerns
